Create form to main_layout
Create the form for the query request in Model and Algorithm page.
This form is stored in the main_layout they are the same in
ML page and Algorithm page.

The "block" in ml_demo.html and algorithm_demo.html are also changed
from {% block content %} to {% block form_action %}