Added Alert + Simple Query Pre-Processing
Added a simple pre-processing function called
"preprocess_text_data()" in app.py that lowercase the
query and apply regex r"\b[a-z]+\b" and "re.findall()"
to that query.

That processed query then displayed to the page in the
form of an alert (the search bar will also be prefilled
with the previous query input).