Implemented TF-IDF (Cosine Similarity Algorithm)
Applied the TF-IDF (Cosine Similarity Algorithm) to the
"Algorithm" page.

This algorithm will find the first discussion that contains the search
query and proceeds to display all the discussions that are related
to that discussion (ordered to more relevant to less relevant).

The threshold for Cosine Similarity Value is 0.1 (any discussion
with Cosine Similarity Value < 0.1 won't be shown).